requirements.txt
Pillow==9.5.0
PyMuPDF==1.21.1